feat(backend): integration tests and RegisterThemeDto

+++ backend/src/main/java/pt/ulisboa/tecnico/socialsoftware/humanaethica/theme/ThemeService.java
@@ -8,6 +8,7 @@ import pt.ulisboa.tecnico.socialsoftware.humanaethica.institution.repository.Ins
 import pt.ulisboa.tecnico.socialsoftware.humanaethica.theme.domain.Theme;
 import pt.ulisboa.tecnico.socialsoftware.humanaethica.exceptions.HEException;
 import pt.ulisboa.tecnico.socialsoftware.humanaethica.theme.dto.ThemeDto;
+import pt.ulisboa.tecnico.socialsoftware.humanaethica.theme.dto.RegisterThemeDto;
 import pt.ulisboa.tecnico.socialsoftware.humanaethica.theme.repository.ThemeRepository;
 import pt.ulisboa.tecnico.socialsoftware.humanaethica.user.repository.UserRepository;
 import static pt.ulisboa.tecnico.socialsoftware.humanaethica.exceptions.ErrorMessage.*;
@@ -36,21 +37,26 @@ public class ThemeService {
     }
 
     @Transactional(isolation = Isolation.READ_COMMITTED)
-    public void registerTheme(ThemeDto themeDto) {
+    public Theme registerTheme(RegisterThemeDto registerThemeDto) {
 
-        if (themeDto.getName() == null || themeDto.getName().trim().length() == 0) {
-            throw new HEException(INVALID_THEME_NAME, themeDto.getName());
+        if (registerThemeDto.getName() == null) {
+            throw new HEException(INVALID_THEME_NAME, registerThemeDto.getName());
         }
 
-        Theme theme = new Theme(themeDto.getName());
+        Theme theme = new Theme(registerThemeDto.getName());
         themeRepository.save(theme);
+        return theme;
     }
 
     @Transactional(isolation = Isolation.READ_COMMITTED)
     public List<ThemeDto> deleteTheme(int themeId) {
         Theme theme = themeRepository.findById(themeId).orElseThrow(() -> new HEException(THEME_NOT_FOUND,Integer.toString(themeId)));
-
-        theme.delete();
-        return getThemes();
+        if (theme.getInstitutions().size() == 0) {
+            theme.delete();
+            return getThemes();
+        }
+        else{
+            throw new HEException(THEME_HAS_INSTITUTIONS, theme.getName());
+        }
     }
 }
